fix: modal show a riddle
