+++ time-react/frontend/src/components/Gameplay/Gameplay.jsx
@@ -3,7 +3,6 @@ import React, { useState, useEffect } from "react";
 import { faGem } from "@fortawesome/free-solid-svg-icons"; */
 import "./Gameplay.css";
 import Button from "../Button/Button";
-import ShowRiddle from "../ShowRiddle/ShowRiddle";
 import userPic from "../../assets/userpic.svg";
 import placeAlienOnGrid from "../../../src/utilits/placeAlienOnGrid";
 import alienImage from "../../assets/alienbob.png";
@@ -17,57 +16,24 @@ function Gameplay({ onLogOut }) {
   const [currentFieldIndex, setCurrentFieldIndex] = useState(0);
   const [riddle, setRiddle] = useState(null);
   const [userAnswer, setUserAnswer] = useState("");
+  const [isModalOpen, setIsModalOpen] = useState(false);//modal is closed by default
+  const openModal = () => setIsModalOpen(true);
+  const closeModal = () => setIsModalOpen(false);
   const [rewards, setRewards] = useState([]);
 
   // Initialize the avatar at the starting grid
   useEffect(() => {
     const startingFieldId = `${gameplayFields[0]}`; // Get the starting grid ID
     placeAlienOnGrid(startingFieldId, alienImage); // Place the alien on the grid
-    fetchRiddle(); // Fetch the first riddle
   }, []);
-
-  // Fetch a new riddle from the API
-  const fetchRiddle = async () => {
-    try {
-      const response = await fetch("https://riddles-api.vercel.app/random");
-      const data = await response.json();
-      console.log("Fetched Riddle:", data); // Debugging the response
-      setRiddle(data); // Set the full API response
-    } catch (error) {
-      console.error("Failed to fetch riddle:", error);
-      setRiddle({
-        riddle: "Error fetching riddle. Please try again later.",
-        answer: "",
-      });
-    }
-  };
 
   const handleMoveNext = () => {
     const nextIndex = (currentFieldIndex + 1) % gameplayFields.length;
     const nextFieldId = `${gameplayFields[nextIndex]}`;
     placeAlienOnGrid(nextFieldId, alienImage); // Move the alien to the next grid
     setCurrentFieldIndex(nextIndex);
-    fetchRiddle(); // Fetch a new riddle when moving to the next stage
     setUserAnswer(""); // Reset the answer input
   };
-
-  const handleAnswerSubmit = () => {
-    if (
-      userAnswer.trim().toLowerCase() === riddle?.answer?.trim().toLowerCase()
-    ) {
-      alert("Correct! You earned a gem.");
-      setRewards([...rewards, "Gem"]); // Add gem to rewards
-      handleMoveNext(); // Automatically move the avatar to the next grid
-    } else {
-      alert("Incorrect answer. Try again.");
-    }
-  };
-
-  //Modal block starts here
-  const [isModalOpen, setIsModalOpen] = useState(false);//modal is closed by default
-  const openModal = () => setIsModalOpen(true);
-  const closeModal = () => setIsModalOpen(false);
-  //Modal block ends here
 
   return (
     <main>
@@ -111,28 +77,12 @@ function Gameplay({ onLogOut }) {
             />
             <Modal
               isOpen={isModalOpen}
-              onClose={closeModal}>
-              <div className="riddle-section">
-                {/* Your riddles go here */
-                  'Test'
-                }
-              </div>
-            </Modal>
-          </div>
-          <div className="riddle-section">
-            <ShowRiddle riddle={riddle} />
-            <input
-              type="text"
-              value={userAnswer}
-              onChange={(e) => setUserAnswer(e.target.value)}
-              placeholder="Enter your answer"
-            />
-            <Button
-              text="Submit Answer"
-              onClick={handleAnswerSubmit}
-              data-role="secondary"
+              onClose={closeModal}
+              handleMoveNext={handleMoveNext}
+              setRewards={setRewards}
             />
           </div>
+
           <Button text="Exit" onClick={onLogOut} data-role="primary" />
           <div id="settings">
             <a href="#">
